requirements.txt
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-storages==1.10.1
filelock==3.0.12
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.34
whitenoise==5.2.0
Collectfast==2.2.0
django-tinymce==3.1.0
django-js-asset==1.2.2
typing-extensions==3.7.4.3